Added input.txt, list of commands to complete the game.
Modified the vm to exit when stdin reaches EOF. It's possible to
complete the game now by redirecting stdin from the input.txt file.

Fixed teleporter patch code to remove the check for non-zero R7.

Added stats (maximum stack size, number of pages modified, and number
of instructions executed since the previous turn) for determining the
feasibility of a 6502 port.